third_party/blink/web_tests/external/wpt/css/css-anchor-position/anchor-scope-shadow-names.html

<!DOCTYPE html>
<title>CSS Anchor Positioning: anchor-scope name match is tree-scoped</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-position-1/#anchor-scope">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/10526">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #host::part(scope) {
    /* This should have no effect, because --a is specified in a different
       tree-scope than the --a inside the shadow. */
    anchor-scope: --a;
  }
</style>
<div id=host>
  <template shadowrootmode=open>
    <style>
      .anchored {
        background: coral;
        position: absolute;
        top: anchor(bottom, 1px);
        position-anchor: --a;
        width: 5px;
        height: 5px;
      }
      .anchor {
        background: skyblue;
        height: 10px;
        anchor-name: --a;
      }
      .cb {
        position: relative;
        width: 200px;
        height: 200px;
        border: 1px solid black;
      }
      .scope {
        anchor-scope: --a;
      }
    </style>
    <div class=cb>
      <div class=anchor></div>
      <div part=scope>
        <div class=anchored></div>
      </div>
    </div>
  </template>
</div>
<script>
  test((t) => {
    let e = host.shadowRoot.querySelector('.anchored');
    assert_equals(getComputedStyle(e).top, '10px');
  }, 'anchor-scope matches tree-scoped names');
</script>
